program for odd or even and multiple of 4 or not
Ask the user for a number. Depending on whether the number is even or odd, print out an appropriate message to the user. 
Extras:
If the number is a multiple of 4, print out a different message.

Ask the user for two numbers: one number to check and one number to divide by, If 2nd divides evenly into 1st, tell that to the user. If not, print a different appropriate message.