Program to guess a number b/w 1 to 9 in 5 chances
A program where a random number between 1 to 9 gets generated and ask user to guess that given number in 5 no. of chances. If the user can guess the number little away from the exact number it displays "your guess is too high" and if it is little less than the exact number it displays "your guess is too low" till you guess the exact the number in total 5 chances.