Program to reverse strings like short sentence
Write a program (using functions!) that asks the user for a long string containing multiple words. Print back to the user the same string, except with the words in backwards order. For example, say I type the string:

  My name is Aditya
Then I would see the string:
Aditya is name My
shown back to me.